Try a couple alternate ideas
transpose during parse: This was harder to get right, but I liked it
because it avoided using Vec::insert. This ended up being a little bit
quicker, though I'm not sure if it's because it avoids Vec::insert or if
it's because it does fewer u8 -> u32 conversions while going from ASCII
input to chars.

"slices": avoid an intermediate Vec in part 2. The reduced allocations
did make this a little bit faster.

Both of these took a little more effort to get right.